+++ routes/web.php
@@ -71,6 +71,14 @@ Route::post('admin/nieuwsbrief/opslaan', 'Backend\NewsLetterController@store')->
 Route::get('admin/nieuwsbrief/zend/{slug}', 'Backend\NewsLetterController@send')->name('admin.nieuwsbrief.zend');
 Route::patch('admin/nieuwsbrief/{slug}', 'Backend\NewsLetterController@update')->name('admin.nieuwsbrief.update');
 
+// Category Management routes
+Route::get('/admin/categorieen', 'Backend\TagsController@index')->name('admin.categories.index');
+Route::get('/admin/categorieen/nieuw', 'Backend\TagsController@create')->name('admin.categories.create');
+Route::get('/admin/categorieen/wijzigen/{id}', 'Backend\TagsController@edit')->name('admin.categories.edit');
+Route::get('/admin/categorieen/verwijder/{id}', 'Backend\TagsController@destroy')->name('admin.categories.delete');
+Route::post('/admin/categorieen/opslaan', 'Backend\TagsController@store')->name('admin.categories.store');
+Route::patch('/admin/categorieen/wijzig/{id}', 'Backend\TagsController@update')->name('admin.categories.update');
+
 // Contact route 
 Route::post('/contact', 'Frontend\ContactController@send')->name('contact.send');