+++ CHANGELOG.md
@@ -8,4 +8,5 @@ All notable changes to `misfits-BE/activisme.be` will be documented in this file
 - Fix #71 | Global user variable.
 - Fix #73 | Datum notitie in formulier is verkeerd
 - Fix #75 | Implementatie `nunomaduro/collision` package.
+- Fix #77 | Beheers console voor de categories v/d nieuwsberichten. 
 - Fix #78 | Implementatie dropdown voor nieuws categorieen.